analysis/requirements.txt
python-dotenv
wget
python-frontmatter